# api-specification.yaml
openapi: 3.0.3
info:
  title: Mini Commerce MSA API
  description: |
    한국형 이커머스 플랫폼을 위한 마이크로서비스 아키텍처 API 명세서
    
    ## 아키텍처 개요
    - **Frontend**: React + TypeScript + Tailwind CSS
    - **API Gateway**: Node.js + Express (포트 8080)
    - **마이크로서비스**: Node.js, Python/FastAPI, Go, Java/Spring Boot
    - **데이터베이스**: PostgreSQL + Redis
    - **인프라**: Kubernetes + Istio Service Mesh
    
    ## 인증
    - JWT (JSON Web Token) 기반 인증
    - Authorization: Bearer {token}
    
    ## 서비스 포트
    - Auth Service: 3001
    - Catalog Service: 3002 (내부 8080)
    - Cart Service: 3003
    - Order Service: 3004 (내부 8080)
    - Inventory Service: 3005
    - Payment Service: 3006
    - Notification Service: 3007
    - Gateway: 8080
    
  version: 1.0.0
  contact:
    name: Mini Commerce Team
    email: dev@minicommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.minicommerce.com
    description: Production server
  - url: https://staging-api.minicommerce.com
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/catalog/products:
    get:
      tags:
        - Catalog
      summary: 상품 목록 조회
      description: 상품 목록을 페이지네이션과 함께 조회합니다.
      parameters:
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: 페이지당 항목 수
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: category
          in: query
          description: 카테고리 필터
          schema:
            type: string
        - name: search
          in: query
          description: 검색 키워드
          schema:
            type: string
        - name: sort
          in: query
          description: 정렬 기준
          schema:
            type: string
            enum: [price_asc, price_desc, name_asc, name_desc, created_at_desc]
            default: created_at_desc
        - name: min_price
          in: query
          description: 최소 가격
          schema:
            type: number
        - name: max_price
          in: query
          description: 최대 가격
          schema:
            type: number
      responses:
        '200':
          description: 상품 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/catalog/products/{productId}:
    get:
      tags:
        - Catalog
      summary: 상품 상세 조회
      description: 특정 상품의 상세 정보를 조회합니다.
      parameters:
        - name: productId
          in: path
          required: true
          description: 상품 ID
          schema:
            type: integer
      responses:
        '200':
          description: 상품 상세 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ProductDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "스마트폰 케이스"
        description:
          type: string
          example: "고품질 실리콘 케이스"
        price:
          type: number
          format: decimal
          example: 15000.00
        original_price:
          type: number
          format: decimal
          example: 20000.00
        discount_rate:
          type: number
          format: decimal
          example: 25.0
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: "전자제품"
        stock_quantity:
          type: integer
          example: 100
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/image1.jpg", "https://example.com/image2.jpg"]
        tags:
          type: array
          items:
            type: string
          example: ["케이스", "스마트폰", "보호"]
        is_featured:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            specifications:
              type: object
              example:
                material: "실리콘"
                color: "투명"
                size: "6.1인치"
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            related_products:
              type: array
              items:
                $ref: '#/components/schemas/Product'
...
    Review:
      type: object
      properties:
        id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        user_name:
          type: string
          example: "홍길동"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        title:
          type: string
          example: "정말 좋은 제품입니다"
        content:
          type: string
          example: "품질이 우수하고 가격도 합리적입니다."
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/review1.jpg"]
        is_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
  responses:
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "요청한 리소스를 찾을 수 없습니다."
...